Enforce SoundFreq config value to be a float (or compiler will take it as a double) Add Debugger and ReleaseDebugger configuration to VS2019 project file

+++ Src/OSD/SDL/Main.cpp
@@ -1405,7 +1405,7 @@ static Util::Config::Node DefaultConfig()
   config.Set("BalanceLeftRight", "0");
   config.Set("BalanceFrontRear", "0");
   config.Set("NbSoundChannels", "4");
-  config.Set("SoundFreq", 57.6); // 60.0? 57.524160f?
+  config.Set("SoundFreq", (float)57.6); // 60.0f? 57.524160f?
   // CDSB
   config.Set("EmulateDSB", true);
   config.Set("SoundVolume", "100");
